Some fixes based on review comments. More coming
